handle partial service failures
when there is overlap in the instances serving a roomset, try all the
possibilities so as to tolerate the situation where some but not all of
the instances serving the needed roomset are down.

this needs affinity brought back into the picture now. if we have it
always make its first attempt with the instance that served the roomset
for this client last time, then it will stick with the same instance as
long as it is around, which will help with the ordering-based race that
already exists, and will minimize misses when the first instance for a
roomset *stays* down.